Add badger quick login
